### 2025-08-26 #259 @adejumoridwan
pdhs - A package for managing and analyzing Demographic and Health Survey (DHS) data
## What is this?

To make the review process easier and to help you maintain your package, pyOpenSci requires testing, continuous integration, and full documentation for packages submitted for review (read more [here][PackagingGuide]. We want to make this as easy as possible for newcomers, so if you'd like some help in prepping your package you can submit this request and we'll assign someone to help you out!

Before submitting a request, check out our [Packaging Guide][PackagingGuide]. The answers to your questions might be there already.

## Package Info
Submitting Author: Adejumo Ridwan Suleiman (@learndataxyz)
Repository Link (if existing): https://github.com/learndataxyz/pdhs

---

```
Access to high-quality, structured, and timely demographic and health data is essential for researchers, policymakers, and public health professionals. The [Demographic and Health Surveys (DHS) Program](https://www.dhsprogram.com) provides a rich repository of standardized datasets across countries and years. However, accessing and using this data programmatically can be cumbersome due to inconsistencies in interfaces, authentication requirements, and data formatting.

The `pdhs` Python library aims to streamline and simplify interaction with the DHS API. It offers an intuitive, well-documented, and Pythonic interface for querying, retrieving, and managing DHS datasets. By abstracting low-level API complexities, `pdhs` allows users to focus on analysis and application rather than on data wrangling. It supports reproducible research, integrates smoothly with common data science workflows (e.g., pandas, numpy, matplotlib), and promotes broader usage of DHS data in academic, development, and policy contexts.

> In short, `pdhs` bridges the gap between powerful public data and the tools needed to derive meaningful insights from it.

---

`pdhs` is a package for managing and analyzing [Demographic and Health Survey (DHS)](https://www.dhsprogram.com) data. It provides functionality to:

1. Access standard indicator data (via [DHS STATcompiler](https://www.statcompiler.com)) using the [DHS API](https://api.dhsprogram.com/).
2. Identify surveys and datasets relevant to specific analyses.
3. Download survey datasets from the [DHS website](https://dhsprogram.com/data/available-datasets.cfm).
4. Load datasets and associated metadata into Python.
5. Extract variables and combine datasets for pooled multi-survey analyses.

```

- If you selected Unsure/Other, briefly explain:

## How Can We Help?

- Indicate how we can help you prepare your package (select all that apply):
	- [x] Creating a test suite
	- [x] Enabling Continuous Integration
	- [ ] Adding a specific feature
	- [x] Improving Documentation
	- [ ] General Assistance
	- [ ] Other

- Explain briefly below if there's something specific you'd like help with:
