### 2025-08-29 #259 @adejumoridwan
@eliotwrobson thank you very much. I would like to submit my package for review. I have set up some tests, but I would like to know if it is sufficient.
I have not set up continous integration though.

Also what are the things to be improved to consider my package for a review.
