### 2025-09-10 #259 @eliotwrobson
@adejumoridwan apologies for the delay, I think adding continuous integration that includes test coverage would definitely be a prerequisite for review. Handling a few other things for now but can make another pass later with more thorough comments.
